# docs/openapi/management.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Heidi Management API",
    "description" : "Administrative API for Cockpit and management automation. This is not the API used to run wallet processes.",
    "version" : "1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/management/v1/integrations/update/{id}" : {
      "put" : {
        "tags" : [ "integration-controller" ],
        "operationId" : "updateIntegration",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IntegrationPayload"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IntegrationDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      }
    },
...
    "/management/v1/user/me" : {
      "get" : {
        "tags" : [ "user-controller" ],
        "summary" : "Get user profile",
        "operationId" : "getUserProfile",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserProfile"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "user-controller" ],
        "summary" : "Update user profile",
        "operationId" : "updateUserProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      }
    },
...
    "/management/v1/integrations/create" : {
      "post" : {
        "tags" : [ "integration-controller" ],
        "operationId" : "createIntegration",
        "parameters" : [ {
          "name" : "tenantId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IntegrationPayload"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IntegrationDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      }
    },
...
    "/management/v1/proof-schemas/{id}" : {
      "get" : {
        "tags" : [ "proof-scheme-controller" ],
        "operationId" : "getProofScheme",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProofSchemeDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      }
    },
...
    "/management/v1/integrations/{id}" : {
      "get" : {
        "tags" : [ "integration-controller" ],
        "operationId" : "getIntegration",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IntegrationDetail"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      }
    },
    "/management/v1/integrations/overview" : {
      "get" : {
        "tags" : [ "integration-controller" ],
        "operationId" : "getAllIntegrations",
        "parameters" : [ {
          "name" : "isPublic",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IntegrationOverview"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      }
    },
    "/management/v1/integrations/credentials" : {
      "get" : {
        "tags" : [ "integration-controller" ],
        "operationId" : "getAuthorizedCredentials",
        "parameters" : [ {
          "name" : "X-API-KEY",
          "in" : "header",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/IntegrationAuthorizedCredentialsResponse"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      }
    },
...
    "/management/v1/integrations/delete/{id}" : {
      "delete" : {
        "tags" : [ "integration-controller" ],
        "operationId" : "deleteIntegration",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "security" : [ {
          "platformSession" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TrustedAuthorityQuery" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "minLength" : 1
          },
          "values" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "minLength" : 1
            },
            "minItems" : 1
          },
          "credentialId" : {
            "type" : "string"
          }
        },
        "required" : [ "type", "values" ]
      },
      "VerifierInfo" : {
        "type" : "object",
        "properties" : {
          "format" : {
            "type" : "string",
            "minLength" : 1
          },
          "data" : {
            "type" : "string",
            "minLength" : 1
          },
          "credentialIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "minLength" : 1
            }
          }
        },
        "required" : [ "data", "format" ]
      },
...
      "IntegrationPayload" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "credentialIdentifiers" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "scopes" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "verify", "issue" ]
            }
          },
          "isPublic" : {
            "type" : "boolean"
          }
        }
      },
      "IntegrationDetail" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "displayName" : {
            "type" : "string"
          },
          "credentialIdentifiers" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "scopes" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "verify", "issue" ]
            }
          },
          "apiKey" : {
            "type" : "string"
          },
          "tenantId" : {
            "type" : "string"
          },
          "isPublic" : {
            "type" : "boolean"
          }
        }
      },
...
      "CredentialSchemeAttribute" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "name" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string",
            "enum" : [ "STRING", "LINK", "FILE_DOWNLOAD", "MAIL", "PHONE", "NUMBER", "BOOLEAN", "DATE", "TIME", "DATETIME", "LOCATION", "DATEOFBIRTH", "IMAGE", "OTHER" ]
          },
          "isArray" : {
            "type" : "boolean"
          },
          "isSensitive" : {
            "type" : "boolean"
          },
          "isDisclosable" : {
            "type" : "boolean"
          },
          "displayName" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "attributeNameOverrides" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "id", "name", "type" ]
      },
...
      "IssuerSettings" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "issuerKeyType" : {
            "type" : "string",
            "enum" : [ "HARDWARE_BIOMETRIC_AUTH", "SOFTWARE_NO_AUTH", "REMOTE_HSM_PIN_AUTH", "NO_KEY_CLAIM_BINDING" ]
          },
          "doctype" : {
            "type" : "string"
          },
          "namespace" : {
            "type" : "string"
          },
          "vct" : {
            "type" : "string"
          },
          "supportedCredentialTypes" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "SD_JWT", "MSO_MDOC", "ZKP_VC", "W3C_VCDM", "OPENBADGES" ]
            },
            "uniqueItems" : true
          },
          "issClaimOverride" : {
            "type" : "string"
          },
          "kidOverride" : {
            "type" : "string"
          },
          "bbsCredentialType" : {
            "type" : "string"
          },
          "defaultTrustSystem" : {
            "type" : "string",
            "enum" : [ "Switzerland", "EUDI", "Custom", "OIDF", "Default" ]
          },
          "signingKeyIds" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          },
          "statusListId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "credentialOfferType" : {
            "type" : "string",
            "enum" : [ "URI", "VALUE" ]
          },
          "issuanceProfileId" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "id", "issuanceProfileId", "issuerKeyType" ]
      },
...
      "UserProfileRequest" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "type" : "string"
          },
          "thumbnail" : {
            "type" : "string"
          }
        }
      },
...
      "UserProfile" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "tenantId" : {
            "type" : "string"
          },
          "role" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "SUPER_ADMIN", "ADMIN", "MANAGER", "EDITOR", "OPERATOR", "DEVELOPER" ]
            }
          },
          "displayName" : {
            "type" : "string"
          },
          "image" : {
            "type" : "string"
          }
        },
        "required" : [ "email", "role", "tenantId", "userId" ]
      },
...
      "IssuerDefinition" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "logo" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "id", "logo", "slug" ]
      },
      "ProofSchemeDetail" : {
        "type" : "object",
        "properties" : {
          "uuid" : {
            "type" : "string",
            "format" : "uuid"
          },
          "title" : {
            "type" : "string"
          },
          "purpose" : {
            "type" : "string"
          },
          "presentationProfileId" : {
            "type" : "string",
            "minLength" : 1
          },
          "validationLogic" : {
            "type" : "string"
          },
          "validationMode" : {
            "type" : "string",
            "enum" : [ "DISABLED", "ENABLED", "ENFORCED" ]
          },
          "redirectUri" : {
            "type" : "string"
          },
          "tenantId" : {
            "type" : "string"
          },
          "verifierIdentity" : {
            "$ref" : "#/components/schemas/IssuerDefinition"
          },
          "verifierIdentityOverridden" : {
            "type" : "boolean"
          },
          "verifierSigningKeyId" : {
            "type" : "string"
          },
          "proofSigningProviderId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "verifierTrustSystem" : {
            "type" : "string",
            "enum" : [ "Switzerland", "EUDI", "Custom", "OIDF", "Default" ]
          },
          "verifierClientIdScheme" : {
            "type" : "string",
            "enum" : [ "x509_san_dns", "x509_hash", "decentralized_identifier", "openid_federation" ]
          },
          "verifierClientIdSchemeOverridden" : {
            "type" : "boolean"
          },
          "registrationCertificate" : {
            "type" : "string"
          },
          "swissIdentityStatement" : {
            "type" : "string"
          },
          "swissVerificationQueryStatement" : {
            "type" : "string"
          },
          "swissVerificationQueryEnabled" : {
            "type" : "boolean"
          },
          "alwaysIncludeDcqlQuery" : {
            "type" : "boolean"
          },
          "swissProtectedVerificationStatements" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "eudiVerificationTrustAnchors" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "swissVerificationTrustAnchor" : {
            "type" : "string"
          },
          "swissTrustRegistryBaseUrl" : {
            "type" : "string"
          },
          "trustedAuthorities" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TrustedAuthorityQuery"
            }
          },
          "verifierInfos" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/VerifierInfo"
            }
          },
          "createdAt" : {
            "type" : "number"
          },
          "updatedAt" : {
            "type" : "number"
          },
          "credentialSchemes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ReducedCredentialSchemeDetail"
            }
          }
        },
        "required" : [ "createdAt", "credentialSchemes", "presentationProfileId", "purpose", "title", "updatedAt", "uuid", "validationMode" ]
      },
      "ReducedCredentialSchemeDetail" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "credentialIdentifier" : {
            "type" : "string"
          },
          "version" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "issuerDefinition" : {
            "$ref" : "#/components/schemas/IssuerDefinition"
          },
          "attributes" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CredentialSchemeAttribute"
            }
          },
          "issuerSettings" : {
            "$ref" : "#/components/schemas/IssuerSettings"
          }
        },
        "required" : [ "credentialIdentifier", "id", "issuerDefinition", "issuerSettings" ]
      },
...
      "IntegrationOverview" : {
        "type" : "object",
        "properties" : {
          "integrations" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/IntegrationOverviewEntry"
            }
          }
        }
      },
      "IntegrationOverviewEntry" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "displayName" : {
            "type" : "string"
          },
          "credentialIdentifiers" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "scopes" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "verify", "issue" ]
            }
          },
          "tenantId" : {
            "type" : "string"
          },
          "isPublic" : {
            "type" : "boolean"
          }
        }
      },
      "IntegrationAuthorizedCredentialsResponse" : {
        "type" : "object",
        "properties" : {
          "credentialIdentifiers" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "scopes" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "enum" : [ "verify", "issue" ]
            }
          },
          "tenantId" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "platformSession" : {
        "type" : "http",
        "description" : "Authenticated platform session.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}